ARROW-3169: [C++] Break up array-test into multiple compilation units
This makes the code easier to read and also eases up recompilation times
on multicore setups.
