+++ cpp/src/arrow/array-test.cc
@@ -46,48 +46,15 @@
 #include "arrow/util/decimal.h"
 #include "arrow/util/lazy.h"
 
+// This file is compiled together with array-*-test.cc into a single
+// executable array-test.
+
 namespace arrow {
 
 using std::string;
 using std::vector;
 
 using internal::checked_cast;
-
-namespace {
-// used to prevent compiler optimizing away side-effect-less statements
-volatile int throw_away = 0;
-
-// checks if the padding of the buffers of the array is zero
-// also causes valgrind warnings if the padding bytes are uninitialized
-void AssertZeroPadded(const Array& array) {
-  for (const auto& buffer : array.data()->buffers) {
-    if (buffer) {
-      const int64_t padding = buffer->capacity() - buffer->size();
-      std::vector<uint8_t> zeros(padding);
-      ASSERT_EQ(0, memcmp(buffer->data() + buffer->size(), zeros.data(), padding));
-    }
-  }
-}
-
-// Check if the valid buffer bytes are initialized by
-// calling memcmp on them which will cause valgrind warnings otherwise
-void TestInitialized(const Array& array) {
-  for (const auto& buffer : array.data()->buffers) {
-    if (buffer) {
-      std::vector<uint8_t> zeros(buffer->capacity());
-      throw_away = memcmp(buffer->data(), zeros.data(), buffer->size());
-    }
-  }
-}
-
-template <typename BuilderType>
-void FinishAndCheckPadding(BuilderType* builder, std::shared_ptr<Array>* out) {
-  ASSERT_OK(builder->Finish(out));
-  AssertZeroPadded(**out);
-  TestInitialized(**out);
-}
-
-}  // namespace
 
 class TestArray : public ::testing::Test {
  public:
@@ -1000,642 +967,6 @@ TEST(TestPrimitiveAdHoc, FloatingSliceApproxEquals) {
 }
 
 // ----------------------------------------------------------------------
-// String / Binary tests
-
-class TestStringArray : public ::testing::Test {
- public:
-  void SetUp() {
-    chars_ = {'a', 'b', 'b', 'c', 'c', 'c'};
-    offsets_ = {0, 1, 1, 1, 3, 6};
-    valid_bytes_ = {1, 1, 0, 1, 1};
-    expected_ = {"a", "", "", "bb", "ccc"};
-
-    MakeArray();
-  }
-
-  void MakeArray() {
-    length_ = static_cast<int64_t>(offsets_.size()) - 1;
-    value_buf_ = Buffer::Wrap(chars_);
-    offsets_buf_ = Buffer::Wrap(offsets_);
-    ASSERT_OK(BitUtil::BytesToBits(valid_bytes_, default_memory_pool(), &null_bitmap_));
-    null_count_ = CountNulls(valid_bytes_);
-
-    strings_ = std::make_shared<StringArray>(length_, offsets_buf_, value_buf_,
-                                             null_bitmap_, null_count_);
-  }
-
- protected:
-  vector<int32_t> offsets_;
-  vector<char> chars_;
-  vector<uint8_t> valid_bytes_;
-
-  vector<string> expected_;
-
-  std::shared_ptr<Buffer> value_buf_;
-  std::shared_ptr<Buffer> offsets_buf_;
-  std::shared_ptr<Buffer> null_bitmap_;
-
-  int64_t null_count_;
-  int64_t length_;
-
-  std::shared_ptr<StringArray> strings_;
-};
-
-TEST_F(TestStringArray, TestArrayBasics) {
-  ASSERT_EQ(length_, strings_->length());
-  ASSERT_EQ(1, strings_->null_count());
-  ASSERT_OK(ValidateArray(*strings_));
-}
-
-TEST_F(TestStringArray, TestType) {
-  std::shared_ptr<DataType> type = strings_->type();
-
-  ASSERT_EQ(Type::STRING, type->id());
-  ASSERT_EQ(Type::STRING, strings_->type_id());
-}
-
-TEST_F(TestStringArray, TestListFunctions) {
-  int pos = 0;
-  for (size_t i = 0; i < expected_.size(); ++i) {
-    ASSERT_EQ(pos, strings_->value_offset(i));
-    ASSERT_EQ(static_cast<int>(expected_[i].size()), strings_->value_length(i));
-    pos += static_cast<int>(expected_[i].size());
-  }
-}
-
-TEST_F(TestStringArray, TestDestructor) {
-  auto arr = std::make_shared<StringArray>(length_, offsets_buf_, value_buf_,
-                                           null_bitmap_, null_count_);
-}
-
-TEST_F(TestStringArray, TestGetString) {
-  for (size_t i = 0; i < expected_.size(); ++i) {
-    if (valid_bytes_[i] == 0) {
-      ASSERT_TRUE(strings_->IsNull(i));
-    } else {
-      ASSERT_EQ(expected_[i], strings_->GetString(i));
-    }
-  }
-}
-
-TEST_F(TestStringArray, TestEmptyStringComparison) {
-  offsets_ = {0, 0, 0, 0, 0, 0};
-  offsets_buf_ = Buffer::Wrap(offsets_);
-  length_ = static_cast<int64_t>(offsets_.size() - 1);
-
-  auto strings_a = std::make_shared<StringArray>(length_, offsets_buf_, nullptr,
-                                                 null_bitmap_, null_count_);
-  auto strings_b = std::make_shared<StringArray>(length_, offsets_buf_, nullptr,
-                                                 null_bitmap_, null_count_);
-  ASSERT_TRUE(strings_a->Equals(strings_b));
-}
-
-TEST_F(TestStringArray, CompareNullByteSlots) {
-  StringBuilder builder;
-  StringBuilder builder2;
-  StringBuilder builder3;
-
-  ASSERT_OK(builder.Append("foo"));
-  ASSERT_OK(builder2.Append("foo"));
-  ASSERT_OK(builder3.Append("foo"));
-
-  ASSERT_OK(builder.Append("bar"));
-  ASSERT_OK(builder2.AppendNull());
-
-  // same length, but different
-  ASSERT_OK(builder3.Append("xyz"));
-
-  ASSERT_OK(builder.Append("baz"));
-  ASSERT_OK(builder2.Append("baz"));
-  ASSERT_OK(builder3.Append("baz"));
-
-  std::shared_ptr<Array> array, array2, array3;
-  FinishAndCheckPadding(&builder, &array);
-  ASSERT_OK(builder2.Finish(&array2));
-  ASSERT_OK(builder3.Finish(&array3));
-
-  const auto& a1 = checked_cast<const StringArray&>(*array);
-  const auto& a2 = checked_cast<const StringArray&>(*array2);
-  const auto& a3 = checked_cast<const StringArray&>(*array3);
-
-  // The validity bitmaps are the same, the data is different, but the unequal
-  // portion is masked out
-  StringArray equal_array(3, a1.value_offsets(), a1.value_data(), a2.null_bitmap(), 1);
-  StringArray equal_array2(3, a3.value_offsets(), a3.value_data(), a2.null_bitmap(), 1);
-
-  ASSERT_TRUE(equal_array.Equals(equal_array2));
-  ASSERT_TRUE(a2.RangeEquals(equal_array2, 0, 3, 0));
-
-  ASSERT_TRUE(equal_array.Array::Slice(1)->Equals(equal_array2.Array::Slice(1)));
-  ASSERT_TRUE(
-      equal_array.Array::Slice(1)->RangeEquals(0, 2, 0, equal_array2.Array::Slice(1)));
-}
-
-TEST_F(TestStringArray, TestSliceGetString) {
-  StringBuilder builder;
-
-  ASSERT_OK(builder.Append("a"));
-  ASSERT_OK(builder.Append("b"));
-  ASSERT_OK(builder.Append("c"));
-
-  std::shared_ptr<Array> array;
-  ASSERT_OK(builder.Finish(&array));
-  auto s = array->Slice(1, 10);
-  auto arr = std::dynamic_pointer_cast<StringArray>(s);
-  ASSERT_EQ(arr->GetString(0), "b");
-}
-
-// ----------------------------------------------------------------------
-// String builder tests
-
-class TestStringBuilder : public TestBuilder {
- public:
-  void SetUp() {
-    TestBuilder::SetUp();
-    builder_.reset(new StringBuilder(pool_));
-  }
-
-  void Done() {
-    std::shared_ptr<Array> out;
-    FinishAndCheckPadding(builder_.get(), &out);
-
-    result_ = std::dynamic_pointer_cast<StringArray>(out);
-    ASSERT_OK(ValidateArray(*result_));
-  }
-
- protected:
-  std::unique_ptr<StringBuilder> builder_;
-  std::shared_ptr<StringArray> result_;
-};
-
-TEST_F(TestStringBuilder, TestScalarAppend) {
-  vector<string> strings = {"", "bb", "a", "", "ccc"};
-  vector<uint8_t> is_null = {0, 0, 0, 1, 0};
-
-  int N = static_cast<int>(strings.size());
-  int reps = 1000;
-
-  for (int j = 0; j < reps; ++j) {
-    for (int i = 0; i < N; ++i) {
-      if (is_null[i]) {
-        ASSERT_OK(builder_->AppendNull());
-      } else {
-        ASSERT_OK(builder_->Append(strings[i]));
-      }
-    }
-  }
-  Done();
-
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(reps, result_->null_count());
-  ASSERT_EQ(reps * 6, result_->value_data()->size());
-
-  int32_t length;
-  int32_t pos = 0;
-  for (int i = 0; i < N * reps; ++i) {
-    if (is_null[i % N]) {
-      ASSERT_TRUE(result_->IsNull(i));
-    } else {
-      ASSERT_FALSE(result_->IsNull(i));
-      result_->GetValue(i, &length);
-      ASSERT_EQ(pos, result_->value_offset(i));
-      ASSERT_EQ(static_cast<int>(strings[i % N].size()), length);
-      ASSERT_EQ(strings[i % N], result_->GetString(i));
-
-      pos += length;
-    }
-  }
-}
-
-TEST_F(TestStringBuilder, TestAppendVector) {
-  vector<string> strings = {"", "bb", "a", "", "ccc"};
-  vector<uint8_t> valid_bytes = {1, 1, 1, 0, 1};
-
-  int N = static_cast<int>(strings.size());
-  int reps = 1000;
-
-  for (int j = 0; j < reps; ++j) {
-    ASSERT_OK(builder_->AppendValues(strings, valid_bytes.data()));
-  }
-  Done();
-
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(reps, result_->null_count());
-  ASSERT_EQ(reps * 6, result_->value_data()->size());
-
-  int32_t length;
-  int32_t pos = 0;
-  for (int i = 0; i < N * reps; ++i) {
-    if (valid_bytes[i % N]) {
-      ASSERT_FALSE(result_->IsNull(i));
-      result_->GetValue(i, &length);
-      ASSERT_EQ(pos, result_->value_offset(i));
-      ASSERT_EQ(static_cast<int>(strings[i % N].size()), length);
-      ASSERT_EQ(strings[i % N], result_->GetString(i));
-
-      pos += length;
-    } else {
-      ASSERT_TRUE(result_->IsNull(i));
-    }
-  }
-}
-
-TEST_F(TestStringBuilder, TestAppendCStringsWithValidBytes) {
-  const char* strings[] = {nullptr, "aaa", nullptr, "ignored", ""};
-  vector<uint8_t> valid_bytes = {1, 1, 1, 0, 1};
-
-  int N = static_cast<int>(sizeof(strings) / sizeof(strings[0]));
-  int reps = 1000;
-
-  for (int j = 0; j < reps; ++j) {
-    ASSERT_OK(builder_->AppendValues(strings, N, valid_bytes.data()));
-  }
-  Done();
-
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(reps * 3, result_->null_count());
-  ASSERT_EQ(reps * 3, result_->value_data()->size());
-
-  int32_t length;
-  int32_t pos = 0;
-  for (int i = 0; i < N * reps; ++i) {
-    auto string = strings[i % N];
-    if (string && valid_bytes[i % N]) {
-      ASSERT_FALSE(result_->IsNull(i));
-      result_->GetValue(i, &length);
-      ASSERT_EQ(pos, result_->value_offset(i));
-      ASSERT_EQ(static_cast<int32_t>(strlen(string)), length);
-      ASSERT_EQ(strings[i % N], result_->GetString(i));
-
-      pos += length;
-    } else {
-      ASSERT_TRUE(result_->IsNull(i));
-    }
-  }
-}
-
-TEST_F(TestStringBuilder, TestAppendCStringsWithoutValidBytes) {
-  const char* strings[] = {"", "bb", "a", nullptr, "ccc"};
-
-  int N = static_cast<int>(sizeof(strings) / sizeof(strings[0]));
-  int reps = 1000;
-
-  for (int j = 0; j < reps; ++j) {
-    ASSERT_OK(builder_->AppendValues(strings, N));
-  }
-  Done();
-
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(reps, result_->null_count());
-  ASSERT_EQ(reps * 6, result_->value_data()->size());
-
-  int32_t length;
-  int32_t pos = 0;
-  for (int i = 0; i < N * reps; ++i) {
-    if (strings[i % N]) {
-      ASSERT_FALSE(result_->IsNull(i));
-      result_->GetValue(i, &length);
-      ASSERT_EQ(pos, result_->value_offset(i));
-      ASSERT_EQ(static_cast<int32_t>(strlen(strings[i % N])), length);
-      ASSERT_EQ(strings[i % N], result_->GetString(i));
-
-      pos += length;
-    } else {
-      ASSERT_TRUE(result_->IsNull(i));
-    }
-  }
-}
-
-TEST_F(TestStringBuilder, TestZeroLength) {
-  // All buffers are null
-  Done();
-}
-
-// Binary container type
-// TODO(emkornfield) there should be some way to refactor these to avoid code duplicating
-// with String
-class TestBinaryArray : public ::testing::Test {
- public:
-  void SetUp() {
-    chars_ = {'a', 'b', 'b', 'c', 'c', 'c'};
-    offsets_ = {0, 1, 1, 1, 3, 6};
-    valid_bytes_ = {1, 1, 0, 1, 1};
-    expected_ = {"a", "", "", "bb", "ccc"};
-
-    MakeArray();
-  }
-
-  void MakeArray() {
-    length_ = static_cast<int64_t>(offsets_.size() - 1);
-    value_buf_ = Buffer::Wrap(chars_);
-    offsets_buf_ = Buffer::Wrap(offsets_);
-
-    ASSERT_OK(BitUtil::BytesToBits(valid_bytes_, default_memory_pool(), &null_bitmap_));
-    null_count_ = CountNulls(valid_bytes_);
-
-    strings_ = std::make_shared<BinaryArray>(length_, offsets_buf_, value_buf_,
-                                             null_bitmap_, null_count_);
-  }
-
- protected:
-  vector<int32_t> offsets_;
-  vector<char> chars_;
-  vector<uint8_t> valid_bytes_;
-
-  vector<string> expected_;
-
-  std::shared_ptr<Buffer> value_buf_;
-  std::shared_ptr<Buffer> offsets_buf_;
-  std::shared_ptr<Buffer> null_bitmap_;
-
-  int64_t null_count_;
-  int64_t length_;
-
-  std::shared_ptr<BinaryArray> strings_;
-};
-
-TEST_F(TestBinaryArray, TestArrayBasics) {
-  ASSERT_EQ(length_, strings_->length());
-  ASSERT_EQ(1, strings_->null_count());
-  ASSERT_OK(ValidateArray(*strings_));
-}
-
-TEST_F(TestBinaryArray, TestType) {
-  std::shared_ptr<DataType> type = strings_->type();
-
-  ASSERT_EQ(Type::BINARY, type->id());
-  ASSERT_EQ(Type::BINARY, strings_->type_id());
-}
-
-TEST_F(TestBinaryArray, TestListFunctions) {
-  size_t pos = 0;
-  for (size_t i = 0; i < expected_.size(); ++i) {
-    ASSERT_EQ(pos, strings_->value_offset(i));
-    ASSERT_EQ(static_cast<int>(expected_[i].size()), strings_->value_length(i));
-    pos += expected_[i].size();
-  }
-}
-
-TEST_F(TestBinaryArray, TestDestructor) {
-  auto arr = std::make_shared<BinaryArray>(length_, offsets_buf_, value_buf_,
-                                           null_bitmap_, null_count_);
-}
-
-TEST_F(TestBinaryArray, TestGetValue) {
-  for (size_t i = 0; i < expected_.size(); ++i) {
-    if (valid_bytes_[i] == 0) {
-      ASSERT_TRUE(strings_->IsNull(i));
-    } else {
-      ASSERT_FALSE(strings_->IsNull(i));
-      ASSERT_EQ(strings_->GetString(i), expected_[i]);
-    }
-  }
-}
-
-TEST_F(TestBinaryArray, TestNullValuesInitialized) {
-  for (size_t i = 0; i < expected_.size(); ++i) {
-    if (valid_bytes_[i] == 0) {
-      ASSERT_TRUE(strings_->IsNull(i));
-    } else {
-      ASSERT_FALSE(strings_->IsNull(i));
-      ASSERT_EQ(strings_->GetString(i), expected_[i]);
-    }
-  }
-  TestInitialized(*strings_);
-}
-
-TEST_F(TestBinaryArray, TestPaddingZeroed) { AssertZeroPadded(*strings_); }
-
-TEST_F(TestBinaryArray, TestGetString) {
-  for (size_t i = 0; i < expected_.size(); ++i) {
-    if (valid_bytes_[i] == 0) {
-      ASSERT_TRUE(strings_->IsNull(i));
-    } else {
-      std::string val = strings_->GetString(i);
-      ASSERT_EQ(0, std::memcmp(expected_[i].data(), val.c_str(), val.size()));
-    }
-  }
-}
-
-TEST_F(TestBinaryArray, TestEqualsEmptyStrings) {
-  BinaryBuilder builder;
-
-  string empty_string("");
-  for (int i = 0; i < 5; ++i) {
-    ASSERT_OK(builder.Append(empty_string));
-  }
-
-  std::shared_ptr<Array> left_arr;
-  FinishAndCheckPadding(&builder, &left_arr);
-
-  const BinaryArray& left = checked_cast<const BinaryArray&>(*left_arr);
-  std::shared_ptr<Array> right =
-      std::make_shared<BinaryArray>(left.length(), left.value_offsets(), nullptr,
-                                    left.null_bitmap(), left.null_count());
-
-  ASSERT_TRUE(left.Equals(right));
-  ASSERT_TRUE(left.RangeEquals(0, left.length(), 0, right));
-}
-
-class TestBinaryBuilder : public TestBuilder {
- public:
-  void SetUp() {
-    TestBuilder::SetUp();
-    builder_.reset(new BinaryBuilder(pool_));
-  }
-
-  void Done() {
-    std::shared_ptr<Array> out;
-    FinishAndCheckPadding(builder_.get(), &out);
-
-    result_ = std::dynamic_pointer_cast<BinaryArray>(out);
-    ASSERT_OK(ValidateArray(*result_));
-  }
-
- protected:
-  std::unique_ptr<BinaryBuilder> builder_;
-  std::shared_ptr<BinaryArray> result_;
-};
-
-TEST_F(TestBinaryBuilder, TestScalarAppend) {
-  vector<string> strings = {"", "bb", "a", "", "ccc"};
-  vector<uint8_t> is_null = {0, 0, 0, 1, 0};
-
-  int N = static_cast<int>(strings.size());
-  int reps = 10;
-
-  for (int j = 0; j < reps; ++j) {
-    for (int i = 0; i < N; ++i) {
-      if (is_null[i]) {
-        ASSERT_OK(builder_->AppendNull());
-      } else {
-        ASSERT_OK(builder_->Append(strings[i]));
-      }
-    }
-  }
-  Done();
-  ASSERT_OK(ValidateArray(*result_));
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(reps, result_->null_count());
-  ASSERT_EQ(reps * 6, result_->value_data()->size());
-
-  int32_t length;
-  for (int i = 0; i < N * reps; ++i) {
-    if (is_null[i % N]) {
-      ASSERT_TRUE(result_->IsNull(i));
-    } else {
-      ASSERT_FALSE(result_->IsNull(i));
-      const uint8_t* vals = result_->GetValue(i, &length);
-      ASSERT_EQ(static_cast<int>(strings[i % N].size()), length);
-      ASSERT_EQ(0, std::memcmp(vals, strings[i % N].data(), length));
-    }
-  }
-}
-
-TEST_F(TestBinaryBuilder, TestScalarAppendUnsafe) {
-  vector<string> strings = {"", "bb", "a", "", "ccc"};
-  vector<uint8_t> is_null = {0, 0, 0, 1, 0};
-
-  int N = static_cast<int>(strings.size());
-  int reps = 13;
-  int total_length = 0;
-  for (auto&& s : strings) total_length += static_cast<int>(s.size());
-
-  ASSERT_OK(builder_->Reserve(N * reps));
-  ASSERT_OK(builder_->ReserveData(total_length * reps));
-
-  for (int j = 0; j < reps; ++j) {
-    for (int i = 0; i < N; ++i) {
-      if (is_null[i]) {
-        builder_->UnsafeAppendNull();
-      } else {
-        builder_->UnsafeAppend(strings[i]);
-      }
-    }
-  }
-  ASSERT_EQ(builder_->value_data_length(), total_length * reps);
-  Done();
-  ASSERT_OK(ValidateArray(*result_));
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(reps, result_->null_count());
-  ASSERT_EQ(reps * total_length, result_->value_data()->size());
-
-  int32_t length;
-  for (int i = 0; i < N * reps; ++i) {
-    if (is_null[i % N]) {
-      ASSERT_TRUE(result_->IsNull(i));
-    } else {
-      ASSERT_FALSE(result_->IsNull(i));
-      const uint8_t* vals = result_->GetValue(i, &length);
-      ASSERT_EQ(static_cast<int>(strings[i % N].size()), length);
-      ASSERT_EQ(0, std::memcmp(vals, strings[i % N].data(), length));
-    }
-  }
-}
-
-TEST_F(TestBinaryBuilder, TestCapacityReserve) {
-  vector<string> strings = {"aaaaa", "bbbbbbbbbb", "ccccccccccccccc", "dddddddddd"};
-  int N = static_cast<int>(strings.size());
-  int reps = 15;
-  int64_t length = 0;
-  int64_t capacity = 1000;
-  int64_t expected_capacity = BitUtil::RoundUpToMultipleOf64(capacity);
-
-  ASSERT_OK(builder_->ReserveData(capacity));
-
-  ASSERT_EQ(length, builder_->value_data_length());
-  ASSERT_EQ(expected_capacity, builder_->value_data_capacity());
-
-  for (int j = 0; j < reps; ++j) {
-    for (int i = 0; i < N; ++i) {
-      ASSERT_OK(builder_->Append(strings[i]));
-      length += static_cast<int>(strings[i].size());
-
-      ASSERT_EQ(length, builder_->value_data_length());
-      ASSERT_EQ(expected_capacity, builder_->value_data_capacity());
-    }
-  }
-
-  int extra_capacity = 500;
-  expected_capacity = BitUtil::RoundUpToMultipleOf64(length + extra_capacity);
-
-  ASSERT_OK(builder_->ReserveData(extra_capacity));
-
-  ASSERT_EQ(length, builder_->value_data_length());
-  ASSERT_EQ(expected_capacity, builder_->value_data_capacity());
-
-  Done();
-
-  ASSERT_EQ(reps * N, result_->length());
-  ASSERT_EQ(0, result_->null_count());
-  ASSERT_EQ(reps * 40, result_->value_data()->size());
-
-  // Capacity is shrunk after `Finish`
-  ASSERT_EQ(640, result_->value_data()->capacity());
-}
-
-TEST_F(TestBinaryBuilder, TestZeroLength) {
-  // All buffers are null
-  Done();
-}
-
-// ----------------------------------------------------------------------
-// Slice tests
-
-template <typename TYPE>
-void CheckSliceEquality() {
-  using Traits = TypeTraits<TYPE>;
-  using BuilderType = typename Traits::BuilderType;
-
-  BuilderType builder;
-
-  vector<string> strings = {"foo", "", "bar", "baz", "qux", ""};
-  vector<uint8_t> is_null = {0, 1, 0, 1, 0, 0};
-
-  int N = static_cast<int>(strings.size());
-  int reps = 10;
-
-  for (int j = 0; j < reps; ++j) {
-    for (int i = 0; i < N; ++i) {
-      if (is_null[i]) {
-        ASSERT_OK(builder.AppendNull());
-      } else {
-        ASSERT_OK(builder.Append(strings[i]));
-      }
-    }
-  }
-
-  std::shared_ptr<Array> array;
-  FinishAndCheckPadding(&builder, &array);
-
-  std::shared_ptr<Array> slice, slice2;
-
-  slice = array->Slice(5);
-  slice2 = array->Slice(5);
-  ASSERT_EQ(N * reps - 5, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(5, slice->length(), 0, slice));
-
-  // Chained slices
-  slice2 = array->Slice(2)->Slice(3);
-  ASSERT_TRUE(slice->Equals(slice2));
-
-  slice = array->Slice(5, 20);
-  slice2 = array->Slice(5, 20);
-  ASSERT_EQ(20, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(5, 25, 0, slice));
-}
-
-TEST_F(TestBinaryArray, TestSliceEquality) { CheckSliceEquality<BinaryType>(); }
-
-TEST_F(TestStringArray, TestSliceEquality) { CheckSliceEquality<BinaryType>(); }
-
-TEST_F(TestBinaryArray, LengthZeroCtor) { BinaryArray array(0, nullptr, nullptr); }
-
-// ----------------------------------------------------------------------
 // FixedSizeBinary tests
 
 class TestFWBinaryArray : public ::testing::Test {
@@ -2116,1533 +1447,6 @@ TEST_F(TestAdaptiveUIntBuilder, TestAppendNulls) {
   for (unsigned index = 0; index < size; ++index) {
     ASSERT_EQ(result_->IsValid(index), static_cast<bool>(valid_bytes[index]));
   }
-}
-
-// ----------------------------------------------------------------------
-// Dictionary tests
-
-template <typename Type>
-class TestDictionaryBuilder : public TestBuilder {};
-
-typedef ::testing::Types<Int8Type, UInt8Type, Int16Type, UInt16Type, Int32Type,
-                         UInt32Type, Int64Type, UInt64Type, FloatType, DoubleType>
-    PrimitiveDictionaries;
-
-TYPED_TEST_CASE(TestDictionaryBuilder, PrimitiveDictionaries);
-
-TYPED_TEST(TestDictionaryBuilder, Basic) {
-  DictionaryBuilder<TypeParam> builder(default_memory_pool());
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-
-  std::shared_ptr<Array> result;
-  ASSERT_OK(builder.Finish(&result));
-
-  // Build expected data
-  NumericBuilder<TypeParam> dict_builder;
-  ASSERT_OK(dict_builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(dict_builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  std::shared_ptr<Array> dict_array;
-  ASSERT_OK(dict_builder.Finish(&dict_array));
-  auto dtype = std::make_shared<DictionaryType>(int8(), dict_array);
-
-  Int8Builder int_builder;
-  ASSERT_OK(int_builder.Append(0));
-  ASSERT_OK(int_builder.Append(1));
-  ASSERT_OK(int_builder.Append(0));
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TYPED_TEST(TestDictionaryBuilder, ArrayConversion) {
-  NumericBuilder<TypeParam> builder;
-  // DictionaryBuilder<TypeParam> builder;
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-
-  std::shared_ptr<Array> intermediate_result;
-  ASSERT_OK(builder.Finish(&intermediate_result));
-  DictionaryBuilder<TypeParam> dictionary_builder(default_memory_pool());
-  ASSERT_OK(dictionary_builder.AppendArray(*intermediate_result));
-  std::shared_ptr<Array> result;
-  ASSERT_OK(dictionary_builder.Finish(&result));
-
-  // Build expected data
-  NumericBuilder<TypeParam> dict_builder;
-  ASSERT_OK(dict_builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(dict_builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  std::shared_ptr<Array> dict_array;
-  ASSERT_OK(dict_builder.Finish(&dict_array));
-  auto dtype = std::make_shared<DictionaryType>(int8(), dict_array);
-
-  Int8Builder int_builder;
-  ASSERT_OK(int_builder.Append(0));
-  ASSERT_OK(int_builder.Append(1));
-  ASSERT_OK(int_builder.Append(0));
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TYPED_TEST(TestDictionaryBuilder, DoubleTableSize) {
-  using Scalar = typename TypeParam::c_type;
-  // Skip this test for (u)int8
-  if (sizeof(Scalar) > 1) {
-    // Build the dictionary Array
-    DictionaryBuilder<TypeParam> builder(default_memory_pool());
-    // Build expected data
-    NumericBuilder<TypeParam> dict_builder;
-    Int16Builder int_builder;
-
-    // Fill with 1024 different values
-    for (int64_t i = 0; i < 1024; i++) {
-      ASSERT_OK(builder.Append(static_cast<Scalar>(i)));
-      ASSERT_OK(dict_builder.Append(static_cast<Scalar>(i)));
-      ASSERT_OK(int_builder.Append(static_cast<uint16_t>(i)));
-    }
-    // Fill with an already existing value
-    for (int64_t i = 0; i < 1024; i++) {
-      ASSERT_OK(builder.Append(static_cast<Scalar>(1)));
-      ASSERT_OK(int_builder.Append(1));
-    }
-
-    // Finalize result
-    std::shared_ptr<Array> result;
-    FinishAndCheckPadding(&builder, &result);
-
-    // Finalize expected data
-    std::shared_ptr<Array> dict_array;
-    ASSERT_OK(dict_builder.Finish(&dict_array));
-    auto dtype = std::make_shared<DictionaryType>(int16(), dict_array);
-    std::shared_ptr<Array> int_array;
-    ASSERT_OK(int_builder.Finish(&int_array));
-
-    DictionaryArray expected(dtype, int_array);
-    ASSERT_TRUE(expected.Equals(result));
-  }
-}
-
-TYPED_TEST(TestDictionaryBuilder, DeltaDictionary) {
-  DictionaryBuilder<TypeParam> builder(default_memory_pool());
-
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  std::shared_ptr<Array> result;
-  FinishAndCheckPadding(&builder, &result);
-
-  // Build expected data for the initial dictionary
-  NumericBuilder<TypeParam> dict_builder1;
-  ASSERT_OK(dict_builder1.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(dict_builder1.Append(static_cast<typename TypeParam::c_type>(2)));
-  std::shared_ptr<Array> dict_array1;
-  ASSERT_OK(dict_builder1.Finish(&dict_array1));
-  auto dtype1 = std::make_shared<DictionaryType>(int8(), dict_array1);
-
-  Int8Builder int_builder1;
-  ASSERT_OK(int_builder1.Append(0));
-  ASSERT_OK(int_builder1.Append(1));
-  ASSERT_OK(int_builder1.Append(0));
-  ASSERT_OK(int_builder1.Append(1));
-  std::shared_ptr<Array> int_array1;
-  ASSERT_OK(int_builder1.Finish(&int_array1));
-
-  DictionaryArray expected(dtype1, int_array1);
-  ASSERT_TRUE(expected.Equals(result));
-
-  // extend the dictionary builder with new data
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-
-  std::shared_ptr<Array> result_delta;
-  ASSERT_OK(builder.Finish(&result_delta));
-
-  // Build expected data for the delta dictionary
-  NumericBuilder<TypeParam> dict_builder2;
-  ASSERT_OK(dict_builder2.Append(static_cast<typename TypeParam::c_type>(3)));
-  std::shared_ptr<Array> dict_array2;
-  ASSERT_OK(dict_builder2.Finish(&dict_array2));
-  auto dtype2 = std::make_shared<DictionaryType>(int8(), dict_array2);
-
-  Int8Builder int_builder2;
-  ASSERT_OK(int_builder2.Append(1));
-  ASSERT_OK(int_builder2.Append(2));
-  ASSERT_OK(int_builder2.Append(2));
-  ASSERT_OK(int_builder2.Append(0));
-  ASSERT_OK(int_builder2.Append(2));
-  std::shared_ptr<Array> int_array2;
-  ASSERT_OK(int_builder2.Finish(&int_array2));
-
-  DictionaryArray expected_delta(dtype2, int_array2);
-  ASSERT_TRUE(expected_delta.Equals(result_delta));
-}
-
-TYPED_TEST(TestDictionaryBuilder, DoubleDeltaDictionary) {
-  DictionaryBuilder<TypeParam> builder(default_memory_pool());
-
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  std::shared_ptr<Array> result;
-  FinishAndCheckPadding(&builder, &result);
-
-  // Build expected data for the initial dictionary
-  NumericBuilder<TypeParam> dict_builder1;
-  ASSERT_OK(dict_builder1.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(dict_builder1.Append(static_cast<typename TypeParam::c_type>(2)));
-  std::shared_ptr<Array> dict_array1;
-  ASSERT_OK(dict_builder1.Finish(&dict_array1));
-  auto dtype1 = std::make_shared<DictionaryType>(int8(), dict_array1);
-
-  Int8Builder int_builder1;
-  ASSERT_OK(int_builder1.Append(0));
-  ASSERT_OK(int_builder1.Append(1));
-  ASSERT_OK(int_builder1.Append(0));
-  ASSERT_OK(int_builder1.Append(1));
-  std::shared_ptr<Array> int_array1;
-  ASSERT_OK(int_builder1.Finish(&int_array1));
-
-  DictionaryArray expected(dtype1, int_array1);
-  ASSERT_TRUE(expected.Equals(result));
-
-  // extend the dictionary builder with new data
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-
-  std::shared_ptr<Array> result_delta1;
-  ASSERT_OK(builder.Finish(&result_delta1));
-
-  // Build expected data for the delta dictionary
-  NumericBuilder<TypeParam> dict_builder2;
-  ASSERT_OK(dict_builder2.Append(static_cast<typename TypeParam::c_type>(3)));
-  std::shared_ptr<Array> dict_array2;
-  ASSERT_OK(dict_builder2.Finish(&dict_array2));
-  auto dtype2 = std::make_shared<DictionaryType>(int8(), dict_array2);
-
-  Int8Builder int_builder2;
-  ASSERT_OK(int_builder2.Append(1));
-  ASSERT_OK(int_builder2.Append(2));
-  ASSERT_OK(int_builder2.Append(2));
-  ASSERT_OK(int_builder2.Append(0));
-  ASSERT_OK(int_builder2.Append(2));
-  std::shared_ptr<Array> int_array2;
-  ASSERT_OK(int_builder2.Finish(&int_array2));
-
-  DictionaryArray expected_delta1(dtype2, int_array2);
-  ASSERT_TRUE(expected_delta1.Equals(result_delta1));
-
-  // extend the dictionary builder with new data again
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(1)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(2)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(3)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(4)));
-  ASSERT_OK(builder.Append(static_cast<typename TypeParam::c_type>(5)));
-
-  std::shared_ptr<Array> result_delta2;
-  ASSERT_OK(builder.Finish(&result_delta2));
-
-  // Build expected data for the delta dictionary again
-  NumericBuilder<TypeParam> dict_builder3;
-  ASSERT_OK(dict_builder3.Append(static_cast<typename TypeParam::c_type>(4)));
-  ASSERT_OK(dict_builder3.Append(static_cast<typename TypeParam::c_type>(5)));
-  std::shared_ptr<Array> dict_array3;
-  ASSERT_OK(dict_builder3.Finish(&dict_array3));
-  auto dtype3 = std::make_shared<DictionaryType>(int8(), dict_array3);
-
-  Int8Builder int_builder3;
-  ASSERT_OK(int_builder3.Append(0));
-  ASSERT_OK(int_builder3.Append(1));
-  ASSERT_OK(int_builder3.Append(2));
-  ASSERT_OK(int_builder3.Append(3));
-  ASSERT_OK(int_builder3.Append(4));
-  std::shared_ptr<Array> int_array3;
-  ASSERT_OK(int_builder3.Finish(&int_array3));
-
-  DictionaryArray expected_delta2(dtype3, int_array3);
-  ASSERT_TRUE(expected_delta2.Equals(result_delta2));
-}
-
-TEST(TestStringDictionaryBuilder, Basic) {
-  // Build the dictionary Array
-  StringDictionaryBuilder builder(default_memory_pool());
-  ASSERT_OK(builder.Append("test"));
-  ASSERT_OK(builder.Append("test2"));
-  ASSERT_OK(builder.Append("test"));
-
-  std::shared_ptr<Array> result;
-  ASSERT_OK(builder.Finish(&result));
-
-  // Build expected data
-  StringBuilder str_builder;
-  ASSERT_OK(str_builder.Append("test"));
-  ASSERT_OK(str_builder.Append("test2"));
-  std::shared_ptr<Array> str_array;
-  ASSERT_OK(str_builder.Finish(&str_array));
-  auto dtype = std::make_shared<DictionaryType>(int8(), str_array);
-
-  Int8Builder int_builder;
-  ASSERT_OK(int_builder.Append(0));
-  ASSERT_OK(int_builder.Append(1));
-  ASSERT_OK(int_builder.Append(0));
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TEST(TestStringDictionaryBuilder, DoubleTableSize) {
-  // Build the dictionary Array
-  StringDictionaryBuilder builder(default_memory_pool());
-  // Build expected data
-  StringBuilder str_builder;
-  Int16Builder int_builder;
-
-  // Fill with 1024 different values
-  for (int64_t i = 0; i < 1024; i++) {
-    std::stringstream ss;
-    ss << "test" << i;
-    ASSERT_OK(builder.Append(ss.str()));
-    ASSERT_OK(str_builder.Append(ss.str()));
-    ASSERT_OK(int_builder.Append(static_cast<uint16_t>(i)));
-  }
-  // Fill with an already existing value
-  for (int64_t i = 0; i < 1024; i++) {
-    ASSERT_OK(builder.Append("test1"));
-    ASSERT_OK(int_builder.Append(1));
-  }
-
-  // Finalize result
-  std::shared_ptr<Array> result;
-  FinishAndCheckPadding(&builder, &result);
-
-  // Finalize expected data
-  std::shared_ptr<Array> str_array;
-  ASSERT_OK(str_builder.Finish(&str_array));
-  auto dtype = std::make_shared<DictionaryType>(int16(), str_array);
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TEST(TestStringDictionaryBuilder, DeltaDictionary) {
-  // Build the dictionary Array
-  StringDictionaryBuilder builder(default_memory_pool());
-  ASSERT_OK(builder.Append("test"));
-  ASSERT_OK(builder.Append("test2"));
-  ASSERT_OK(builder.Append("test"));
-
-  std::shared_ptr<Array> result;
-  ASSERT_OK(builder.Finish(&result));
-
-  // Build expected data
-  StringBuilder str_builder1;
-  ASSERT_OK(str_builder1.Append("test"));
-  ASSERT_OK(str_builder1.Append("test2"));
-  std::shared_ptr<Array> str_array1;
-  ASSERT_OK(str_builder1.Finish(&str_array1));
-  auto dtype1 = std::make_shared<DictionaryType>(int8(), str_array1);
-
-  Int8Builder int_builder1;
-  ASSERT_OK(int_builder1.Append(0));
-  ASSERT_OK(int_builder1.Append(1));
-  ASSERT_OK(int_builder1.Append(0));
-  std::shared_ptr<Array> int_array1;
-  ASSERT_OK(int_builder1.Finish(&int_array1));
-
-  DictionaryArray expected(dtype1, int_array1);
-  ASSERT_TRUE(expected.Equals(result));
-
-  // build a delta dictionary
-  ASSERT_OK(builder.Append("test2"));
-  ASSERT_OK(builder.Append("test3"));
-  ASSERT_OK(builder.Append("test2"));
-
-  std::shared_ptr<Array> result_delta;
-  FinishAndCheckPadding(&builder, &result_delta);
-
-  // Build expected data
-  StringBuilder str_builder2;
-  ASSERT_OK(str_builder2.Append("test3"));
-  std::shared_ptr<Array> str_array2;
-  ASSERT_OK(str_builder2.Finish(&str_array2));
-  auto dtype2 = std::make_shared<DictionaryType>(int8(), str_array2);
-
-  Int8Builder int_builder2;
-  ASSERT_OK(int_builder2.Append(1));
-  ASSERT_OK(int_builder2.Append(2));
-  ASSERT_OK(int_builder2.Append(1));
-  std::shared_ptr<Array> int_array2;
-  ASSERT_OK(int_builder2.Finish(&int_array2));
-
-  DictionaryArray expected_delta(dtype2, int_array2);
-  ASSERT_TRUE(expected_delta.Equals(result_delta));
-}
-
-TEST(TestStringDictionaryBuilder, BigDeltaDictionary) {
-  constexpr int16_t kTestLength = 2048;
-  // Build the dictionary Array
-  StringDictionaryBuilder builder(default_memory_pool());
-
-  StringBuilder str_builder1;
-  Int16Builder int_builder1;
-
-  for (int16_t idx = 0; idx < kTestLength; ++idx) {
-    std::stringstream sstream;
-    sstream << "test" << idx;
-    ASSERT_OK(builder.Append(sstream.str()));
-    ASSERT_OK(str_builder1.Append(sstream.str()));
-    ASSERT_OK(int_builder1.Append(idx));
-  }
-
-  std::shared_ptr<Array> result;
-  FinishAndCheckPadding(&builder, &result);
-
-  std::shared_ptr<Array> str_array1;
-  ASSERT_OK(str_builder1.Finish(&str_array1));
-  auto dtype1 = std::make_shared<DictionaryType>(int16(), str_array1);
-
-  std::shared_ptr<Array> int_array1;
-  ASSERT_OK(int_builder1.Finish(&int_array1));
-
-  DictionaryArray expected(dtype1, int_array1);
-  ASSERT_TRUE(expected.Equals(result));
-
-  // build delta 1
-  StringBuilder str_builder2;
-  Int16Builder int_builder2;
-
-  for (int16_t idx = 0; idx < kTestLength; ++idx) {
-    ASSERT_OK(builder.Append("test1"));
-    ASSERT_OK(int_builder2.Append(1));
-  }
-
-  for (int16_t idx = 0; idx < kTestLength; ++idx) {
-    ASSERT_OK(builder.Append("test_new_value1"));
-    ASSERT_OK(int_builder2.Append(kTestLength));
-  }
-  ASSERT_OK(str_builder2.Append("test_new_value1"));
-
-  std::shared_ptr<Array> result2;
-  ASSERT_OK(builder.Finish(&result2));
-
-  std::shared_ptr<Array> str_array2;
-  ASSERT_OK(str_builder2.Finish(&str_array2));
-  auto dtype2 = std::make_shared<DictionaryType>(int16(), str_array2);
-
-  std::shared_ptr<Array> int_array2;
-  ASSERT_OK(int_builder2.Finish(&int_array2));
-
-  DictionaryArray expected2(dtype2, int_array2);
-  ASSERT_TRUE(expected2.Equals(result2));
-
-  // build delta 2
-  StringBuilder str_builder3;
-  Int16Builder int_builder3;
-
-  for (int16_t idx = 0; idx < kTestLength; ++idx) {
-    ASSERT_OK(builder.Append("test2"));
-    ASSERT_OK(int_builder3.Append(2));
-  }
-
-  for (int16_t idx = 0; idx < kTestLength; ++idx) {
-    ASSERT_OK(builder.Append("test_new_value2"));
-    ASSERT_OK(int_builder3.Append(kTestLength + 1));
-  }
-  ASSERT_OK(str_builder3.Append("test_new_value2"));
-
-  std::shared_ptr<Array> result3;
-  ASSERT_OK(builder.Finish(&result3));
-
-  std::shared_ptr<Array> str_array3;
-  ASSERT_OK(str_builder3.Finish(&str_array3));
-  auto dtype3 = std::make_shared<DictionaryType>(int16(), str_array3);
-
-  std::shared_ptr<Array> int_array3;
-  ASSERT_OK(int_builder3.Finish(&int_array3));
-
-  DictionaryArray expected3(dtype3, int_array3);
-  ASSERT_TRUE(expected3.Equals(result3));
-}
-
-TEST(TestFixedSizeBinaryDictionaryBuilder, Basic) {
-  // Build the dictionary Array
-  DictionaryBuilder<FixedSizeBinaryType> builder(arrow::fixed_size_binary(4),
-                                                 default_memory_pool());
-  std::vector<uint8_t> test{12, 12, 11, 12};
-  std::vector<uint8_t> test2{12, 12, 11, 11};
-  ASSERT_OK(builder.Append(test.data()));
-  ASSERT_OK(builder.Append(test2.data()));
-  ASSERT_OK(builder.Append(test.data()));
-
-  std::shared_ptr<Array> result;
-  FinishAndCheckPadding(&builder, &result);
-
-  // Build expected data
-  FixedSizeBinaryBuilder fsb_builder(arrow::fixed_size_binary(4));
-  ASSERT_OK(fsb_builder.Append(test.data()));
-  ASSERT_OK(fsb_builder.Append(test2.data()));
-  std::shared_ptr<Array> fsb_array;
-  ASSERT_OK(fsb_builder.Finish(&fsb_array));
-  auto dtype = std::make_shared<DictionaryType>(int8(), fsb_array);
-
-  Int8Builder int_builder;
-  ASSERT_OK(int_builder.Append(0));
-  ASSERT_OK(int_builder.Append(1));
-  ASSERT_OK(int_builder.Append(0));
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TEST(TestFixedSizeBinaryDictionaryBuilder, DeltaDictionary) {
-  // Build the dictionary Array
-  DictionaryBuilder<FixedSizeBinaryType> builder(arrow::fixed_size_binary(4),
-                                                 default_memory_pool());
-  std::vector<uint8_t> test{12, 12, 11, 12};
-  std::vector<uint8_t> test2{12, 12, 11, 11};
-  std::vector<uint8_t> test3{12, 12, 11, 10};
-
-  ASSERT_OK(builder.Append(test.data()));
-  ASSERT_OK(builder.Append(test2.data()));
-  ASSERT_OK(builder.Append(test.data()));
-
-  std::shared_ptr<Array> result1;
-  FinishAndCheckPadding(&builder, &result1);
-
-  // Build expected data
-  FixedSizeBinaryBuilder fsb_builder1(arrow::fixed_size_binary(4));
-  ASSERT_OK(fsb_builder1.Append(test.data()));
-  ASSERT_OK(fsb_builder1.Append(test2.data()));
-  std::shared_ptr<Array> fsb_array1;
-  ASSERT_OK(fsb_builder1.Finish(&fsb_array1));
-  auto dtype1 = std::make_shared<DictionaryType>(int8(), fsb_array1);
-
-  Int8Builder int_builder1;
-  ASSERT_OK(int_builder1.Append(0));
-  ASSERT_OK(int_builder1.Append(1));
-  ASSERT_OK(int_builder1.Append(0));
-  std::shared_ptr<Array> int_array1;
-  ASSERT_OK(int_builder1.Finish(&int_array1));
-
-  DictionaryArray expected1(dtype1, int_array1);
-  ASSERT_TRUE(expected1.Equals(result1));
-
-  // build delta dictionary
-  ASSERT_OK(builder.Append(test.data()));
-  ASSERT_OK(builder.Append(test2.data()));
-  ASSERT_OK(builder.Append(test3.data()));
-
-  std::shared_ptr<Array> result2;
-  FinishAndCheckPadding(&builder, &result2);
-
-  // Build expected data
-  FixedSizeBinaryBuilder fsb_builder2(arrow::fixed_size_binary(4));
-  ASSERT_OK(fsb_builder2.Append(test3.data()));
-  std::shared_ptr<Array> fsb_array2;
-  ASSERT_OK(fsb_builder2.Finish(&fsb_array2));
-  auto dtype2 = std::make_shared<DictionaryType>(int8(), fsb_array2);
-
-  Int8Builder int_builder2;
-  ASSERT_OK(int_builder2.Append(0));
-  ASSERT_OK(int_builder2.Append(1));
-  ASSERT_OK(int_builder2.Append(2));
-  std::shared_ptr<Array> int_array2;
-  ASSERT_OK(int_builder2.Finish(&int_array2));
-
-  DictionaryArray expected2(dtype2, int_array2);
-  ASSERT_TRUE(expected2.Equals(result2));
-}
-
-TEST(TestFixedSizeBinaryDictionaryBuilder, DoubleTableSize) {
-  // Build the dictionary Array
-  DictionaryBuilder<FixedSizeBinaryType> builder(arrow::fixed_size_binary(4),
-                                                 default_memory_pool());
-  // Build expected data
-  FixedSizeBinaryBuilder fsb_builder(arrow::fixed_size_binary(4));
-  Int16Builder int_builder;
-
-  // Fill with 1024 different values
-  for (int64_t i = 0; i < 1024; i++) {
-    std::vector<uint8_t> value{12, 12, static_cast<uint8_t>(i / 128),
-                               static_cast<uint8_t>(i % 128)};
-    ASSERT_OK(builder.Append(value.data()));
-    ASSERT_OK(fsb_builder.Append(value.data()));
-    ASSERT_OK(int_builder.Append(static_cast<uint16_t>(i)));
-  }
-  // Fill with an already existing value
-  std::vector<uint8_t> known_value{12, 12, 0, 1};
-  for (int64_t i = 0; i < 1024; i++) {
-    ASSERT_OK(builder.Append(known_value.data()));
-    ASSERT_OK(int_builder.Append(1));
-  }
-
-  // Finalize result
-  std::shared_ptr<Array> result;
-  ASSERT_OK(builder.Finish(&result));
-
-  // Finalize expected data
-  std::shared_ptr<Array> fsb_array;
-  ASSERT_OK(fsb_builder.Finish(&fsb_array));
-  auto dtype = std::make_shared<DictionaryType>(int16(), fsb_array);
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TEST(TestFixedSizeBinaryDictionaryBuilder, InvalidTypeAppend) {
-  // Build the dictionary Array
-  DictionaryBuilder<FixedSizeBinaryType> builder(arrow::fixed_size_binary(4),
-                                                 default_memory_pool());
-  // Build an array with different byte width
-  FixedSizeBinaryBuilder fsb_builder(arrow::fixed_size_binary(5));
-  std::vector<uint8_t> value{100, 1, 1, 1, 1};
-  ASSERT_OK(fsb_builder.Append(value.data()));
-  std::shared_ptr<Array> fsb_array;
-  ASSERT_OK(fsb_builder.Finish(&fsb_array));
-
-  ASSERT_RAISES(Invalid, builder.AppendArray(*fsb_array));
-}
-
-TEST(TestDecimalDictionaryBuilder, Basic) {
-  // Build the dictionary Array
-  const auto& decimal_type = arrow::decimal(2, 0);
-  DictionaryBuilder<FixedSizeBinaryType> builder(decimal_type, default_memory_pool());
-
-  // Test data
-  std::vector<Decimal128> test{12, 12, 11, 12};
-  for (const auto& value : test) {
-    ASSERT_OK(builder.Append(value.ToBytes().data()));
-  }
-
-  std::shared_ptr<Array> result;
-  ASSERT_OK(builder.Finish(&result));
-
-  // Build expected data
-  FixedSizeBinaryBuilder decimal_builder(decimal_type);
-  ASSERT_OK(decimal_builder.Append(Decimal128(12).ToBytes()));
-  ASSERT_OK(decimal_builder.Append(Decimal128(11).ToBytes()));
-
-  std::shared_ptr<Array> decimal_array;
-  ASSERT_OK(decimal_builder.Finish(&decimal_array));
-  auto dtype = arrow::dictionary(int8(), decimal_array);
-
-  Int8Builder int_builder;
-  ASSERT_OK(int_builder.AppendValues({0, 0, 1, 0}));
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-TEST(TestDecimalDictionaryBuilder, DoubleTableSize) {
-  const auto& decimal_type = arrow::decimal(21, 0);
-
-  // Build the dictionary Array
-  DictionaryBuilder<FixedSizeBinaryType> builder(decimal_type, default_memory_pool());
-
-  // Build expected data
-  FixedSizeBinaryBuilder fsb_builder(decimal_type);
-  Int16Builder int_builder;
-
-  // Fill with 1024 different values
-  for (int64_t i = 0; i < 1024; i++) {
-    const uint8_t bytes[] = {0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             0,
-                             12,
-                             12,
-                             static_cast<uint8_t>(i / 128),
-                             static_cast<uint8_t>(i % 128)};
-    ASSERT_OK(builder.Append(bytes));
-    ASSERT_OK(fsb_builder.Append(bytes));
-    ASSERT_OK(int_builder.Append(static_cast<uint16_t>(i)));
-  }
-  // Fill with an already existing value
-  const uint8_t known_value[] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0, 1};
-  for (int64_t i = 0; i < 1024; i++) {
-    ASSERT_OK(builder.Append(known_value));
-    ASSERT_OK(int_builder.Append(1));
-  }
-
-  // Finalize result
-  std::shared_ptr<Array> result;
-  ASSERT_OK(builder.Finish(&result));
-
-  // Finalize expected data
-  std::shared_ptr<Array> fsb_array;
-  ASSERT_OK(fsb_builder.Finish(&fsb_array));
-
-  auto dtype = std::make_shared<DictionaryType>(int16(), fsb_array);
-  std::shared_ptr<Array> int_array;
-  ASSERT_OK(int_builder.Finish(&int_array));
-
-  DictionaryArray expected(dtype, int_array);
-  ASSERT_TRUE(expected.Equals(result));
-}
-
-// ----------------------------------------------------------------------
-// List tests
-
-class TestListArray : public TestBuilder {
- public:
-  void SetUp() {
-    TestBuilder::SetUp();
-
-    value_type_ = int32();
-    type_ = list(value_type_);
-
-    std::unique_ptr<ArrayBuilder> tmp;
-    ASSERT_OK(MakeBuilder(pool_, type_, &tmp));
-    builder_.reset(checked_cast<ListBuilder*>(tmp.release()));
-  }
-
-  void Done() {
-    std::shared_ptr<Array> out;
-    FinishAndCheckPadding(builder_.get(), &out);
-    result_ = std::dynamic_pointer_cast<ListArray>(out);
-  }
-
- protected:
-  std::shared_ptr<DataType> value_type_;
-
-  std::shared_ptr<ListBuilder> builder_;
-  std::shared_ptr<ListArray> result_;
-};
-
-TEST_F(TestListArray, Equality) {
-  Int32Builder* vb = checked_cast<Int32Builder*>(builder_->value_builder());
-
-  std::shared_ptr<Array> array, equal_array, unequal_array;
-  vector<int32_t> equal_offsets = {0, 1, 2, 5, 6, 7, 8, 10};
-  vector<int32_t> equal_values = {1, 2, 3, 4, 5, 2, 2, 2, 5, 6};
-  vector<int32_t> unequal_offsets = {0, 1, 4, 7};
-  vector<int32_t> unequal_values = {1, 2, 2, 2, 3, 4, 5};
-
-  // setup two equal arrays
-  ASSERT_OK(builder_->AppendValues(equal_offsets.data(), equal_offsets.size()));
-  ASSERT_OK(vb->AppendValues(equal_values.data(), equal_values.size()));
-
-  ASSERT_OK(builder_->Finish(&array));
-  ASSERT_OK(builder_->AppendValues(equal_offsets.data(), equal_offsets.size()));
-  ASSERT_OK(vb->AppendValues(equal_values.data(), equal_values.size()));
-
-  ASSERT_OK(builder_->Finish(&equal_array));
-  // now an unequal one
-  ASSERT_OK(builder_->AppendValues(unequal_offsets.data(), unequal_offsets.size()));
-  ASSERT_OK(vb->AppendValues(unequal_values.data(), unequal_values.size()));
-
-  ASSERT_OK(builder_->Finish(&unequal_array));
-
-  // Test array equality
-  EXPECT_TRUE(array->Equals(array));
-  EXPECT_TRUE(array->Equals(equal_array));
-  EXPECT_TRUE(equal_array->Equals(array));
-  EXPECT_FALSE(equal_array->Equals(unequal_array));
-  EXPECT_FALSE(unequal_array->Equals(equal_array));
-
-  // Test range equality
-  EXPECT_TRUE(array->RangeEquals(0, 1, 0, unequal_array));
-  EXPECT_FALSE(array->RangeEquals(0, 2, 0, unequal_array));
-  EXPECT_FALSE(array->RangeEquals(1, 2, 1, unequal_array));
-  EXPECT_TRUE(array->RangeEquals(2, 3, 2, unequal_array));
-
-  // Check with slices, ARROW-33
-  std::shared_ptr<Array> slice, slice2;
-
-  slice = array->Slice(2);
-  slice2 = array->Slice(2);
-  ASSERT_EQ(array->length() - 2, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(2, slice->length(), 0, slice));
-
-  // Chained slices
-  slice2 = array->Slice(1)->Slice(1);
-  ASSERT_TRUE(slice->Equals(slice2));
-
-  slice = array->Slice(1, 4);
-  slice2 = array->Slice(1, 4);
-  ASSERT_EQ(4, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(1, 5, 0, slice));
-}
-
-TEST_F(TestListArray, TestResize) {}
-
-TEST_F(TestListArray, TestFromArrays) {
-  std::shared_ptr<Array> offsets1, offsets2, offsets3, offsets4, values;
-
-  std::vector<bool> offsets_is_valid3 = {true, false, true, true};
-  std::vector<bool> offsets_is_valid4 = {true, true, false, true};
-
-  std::vector<bool> values_is_valid = {true, false, true, true, true, true};
-
-  std::vector<int32_t> offset1_values = {0, 2, 2, 6};
-  std::vector<int32_t> offset2_values = {0, 2, 6, 6};
-
-  std::vector<int8_t> values_values = {0, 1, 2, 3, 4, 5};
-  const int length = 3;
-
-  ArrayFromVector<Int32Type, int32_t>(offset1_values, &offsets1);
-  ArrayFromVector<Int32Type, int32_t>(offset2_values, &offsets2);
-
-  ArrayFromVector<Int32Type, int32_t>(offsets_is_valid3, offset1_values, &offsets3);
-  ArrayFromVector<Int32Type, int32_t>(offsets_is_valid4, offset2_values, &offsets4);
-
-  ArrayFromVector<Int8Type, int8_t>(values_is_valid, values_values, &values);
-
-  auto list_type = list(int8());
-
-  std::shared_ptr<Array> list1, list3, list4;
-  ASSERT_OK(ListArray::FromArrays(*offsets1, *values, pool_, &list1));
-  ASSERT_OK(ListArray::FromArrays(*offsets3, *values, pool_, &list3));
-  ASSERT_OK(ListArray::FromArrays(*offsets4, *values, pool_, &list4));
-
-  ListArray expected1(list_type, length, offsets1->data()->buffers[1], values,
-                      offsets1->data()->buffers[0], 0);
-  AssertArraysEqual(expected1, *list1);
-
-  // Use null bitmap from offsets3, but clean offsets from non-null version
-  ListArray expected3(list_type, length, offsets1->data()->buffers[1], values,
-                      offsets3->data()->buffers[0], 1);
-  AssertArraysEqual(expected3, *list3);
-
-  // Check that the last offset bit is zero
-  ASSERT_FALSE(BitUtil::GetBit(list3->null_bitmap()->data(), length + 1));
-
-  ListArray expected4(list_type, length, offsets2->data()->buffers[1], values,
-                      offsets4->data()->buffers[0], 1);
-  AssertArraysEqual(expected4, *list4);
-
-  // Test failure modes
-
-  std::shared_ptr<Array> tmp;
-
-  // Zero-length offsets
-  ASSERT_RAISES(Invalid,
-                ListArray::FromArrays(*offsets1->Slice(0, 0), *values, pool_, &tmp));
-
-  // Offsets not int32
-  ASSERT_RAISES(Invalid, ListArray::FromArrays(*values, *offsets1, pool_, &tmp));
-}
-
-TEST_F(TestListArray, TestAppendNull) {
-  ASSERT_OK(builder_->AppendNull());
-  ASSERT_OK(builder_->AppendNull());
-
-  Done();
-
-  ASSERT_OK(ValidateArray(*result_));
-  ASSERT_TRUE(result_->IsNull(0));
-  ASSERT_TRUE(result_->IsNull(1));
-
-  ASSERT_EQ(0, result_->raw_value_offsets()[0]);
-  ASSERT_EQ(0, result_->value_offset(1));
-  ASSERT_EQ(0, result_->value_offset(2));
-
-  auto values = result_->values();
-  ASSERT_EQ(0, values->length());
-  // Values buffer should be non-null
-  ASSERT_NE(nullptr, values->data()->buffers[1]);
-}
-
-void ValidateBasicListArray(const ListArray* result, const vector<int32_t>& values,
-                            const vector<uint8_t>& is_valid) {
-  ASSERT_OK(ValidateArray(*result));
-  ASSERT_EQ(1, result->null_count());
-  ASSERT_EQ(0, result->values()->null_count());
-
-  ASSERT_EQ(3, result->length());
-  vector<int32_t> ex_offsets = {0, 3, 3, 7};
-  for (size_t i = 0; i < ex_offsets.size(); ++i) {
-    ASSERT_EQ(ex_offsets[i], result->value_offset(i));
-  }
-
-  for (int i = 0; i < result->length(); ++i) {
-    ASSERT_EQ(is_valid[i] == 0, result->IsNull(i));
-  }
-
-  ASSERT_EQ(7, result->values()->length());
-  auto varr = std::dynamic_pointer_cast<Int32Array>(result->values());
-
-  for (size_t i = 0; i < values.size(); ++i) {
-    ASSERT_EQ(values[i], varr->Value(i));
-  }
-}
-
-TEST_F(TestListArray, TestBasics) {
-  vector<int32_t> values = {0, 1, 2, 3, 4, 5, 6};
-  vector<int> lengths = {3, 0, 4};
-  vector<uint8_t> is_valid = {1, 0, 1};
-
-  Int32Builder* vb = checked_cast<Int32Builder*>(builder_->value_builder());
-
-  ASSERT_OK(builder_->Reserve(lengths.size()));
-  ASSERT_OK(vb->Reserve(values.size()));
-
-  int pos = 0;
-  for (size_t i = 0; i < lengths.size(); ++i) {
-    ASSERT_OK(builder_->Append(is_valid[i] > 0));
-    for (int j = 0; j < lengths[i]; ++j) {
-      ASSERT_OK(vb->Append(values[pos++]));
-    }
-  }
-
-  Done();
-  ValidateBasicListArray(result_.get(), values, is_valid);
-}
-
-TEST_F(TestListArray, BulkAppend) {
-  vector<int32_t> values = {0, 1, 2, 3, 4, 5, 6};
-  vector<int> lengths = {3, 0, 4};
-  vector<uint8_t> is_valid = {1, 0, 1};
-  vector<int32_t> offsets = {0, 3, 3};
-
-  Int32Builder* vb = checked_cast<Int32Builder*>(builder_->value_builder());
-  ASSERT_OK(vb->Reserve(values.size()));
-
-  ASSERT_OK(builder_->AppendValues(offsets.data(), offsets.size(), is_valid.data()));
-  for (int32_t value : values) {
-    ASSERT_OK(vb->Append(value));
-  }
-  Done();
-  ValidateBasicListArray(result_.get(), values, is_valid);
-}
-
-TEST_F(TestListArray, BulkAppendInvalid) {
-  vector<int32_t> values = {0, 1, 2, 3, 4, 5, 6};
-  vector<int> lengths = {3, 0, 4};
-  vector<uint8_t> is_null = {0, 1, 0};
-  vector<uint8_t> is_valid = {1, 0, 1};
-  vector<int32_t> offsets = {0, 2, 4};  // should be 0, 3, 3 given the is_null array
-
-  Int32Builder* vb = checked_cast<Int32Builder*>(builder_->value_builder());
-  ASSERT_OK(vb->Reserve(values.size()));
-
-  ASSERT_OK(builder_->AppendValues(offsets.data(), offsets.size(), is_valid.data()));
-  ASSERT_OK(builder_->AppendValues(offsets.data(), offsets.size(), is_valid.data()));
-  for (int32_t value : values) {
-    ASSERT_OK(vb->Append(value));
-  }
-
-  Done();
-  ASSERT_RAISES(Invalid, ValidateArray(*result_));
-}
-
-TEST_F(TestListArray, TestZeroLength) {
-  // All buffers are null
-  Done();
-  ASSERT_OK(ValidateArray(*result_));
-}
-
-// ----------------------------------------------------------------------
-// DictionaryArray tests
-
-TEST(TestDictionary, Basics) {
-  vector<int32_t> values = {100, 1000, 10000, 100000};
-  std::shared_ptr<Array> dict;
-  ArrayFromVector<Int32Type, int32_t>(values, &dict);
-
-  std::shared_ptr<DictionaryType> type1 =
-      std::dynamic_pointer_cast<DictionaryType>(dictionary(int16(), dict));
-
-  auto type2 =
-      std::dynamic_pointer_cast<DictionaryType>(::arrow::dictionary(int16(), dict, true));
-
-  ASSERT_TRUE(int16()->Equals(type1->index_type()));
-  ASSERT_TRUE(type1->dictionary()->Equals(dict));
-
-  ASSERT_TRUE(int16()->Equals(type2->index_type()));
-  ASSERT_TRUE(type2->dictionary()->Equals(dict));
-
-  ASSERT_EQ("dictionary<values=int32, indices=int16, ordered=0>", type1->ToString());
-  ASSERT_EQ("dictionary<values=int32, indices=int16, ordered=1>", type2->ToString());
-}
-
-TEST(TestDictionary, Equals) {
-  vector<bool> is_valid = {true, true, false, true, true, true};
-
-  std::shared_ptr<Array> dict;
-  vector<string> dict_values = {"foo", "bar", "baz"};
-  ArrayFromVector<StringType, string>(dict_values, &dict);
-  std::shared_ptr<DataType> dict_type = dictionary(int16(), dict);
-
-  std::shared_ptr<Array> dict2;
-  vector<string> dict2_values = {"foo", "bar", "baz", "qux"};
-  ArrayFromVector<StringType, string>(dict2_values, &dict2);
-  std::shared_ptr<DataType> dict2_type = dictionary(int16(), dict2);
-
-  std::shared_ptr<Array> indices;
-  vector<int16_t> indices_values = {1, 2, -1, 0, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(is_valid, indices_values, &indices);
-
-  std::shared_ptr<Array> indices2;
-  vector<int16_t> indices2_values = {1, 2, 0, 0, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(is_valid, indices2_values, &indices2);
-
-  std::shared_ptr<Array> indices3;
-  vector<int16_t> indices3_values = {1, 1, 0, 0, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(is_valid, indices3_values, &indices3);
-
-  auto array = std::make_shared<DictionaryArray>(dict_type, indices);
-  auto array2 = std::make_shared<DictionaryArray>(dict_type, indices2);
-  auto array3 = std::make_shared<DictionaryArray>(dict2_type, indices);
-  auto array4 = std::make_shared<DictionaryArray>(dict_type, indices3);
-
-  ASSERT_TRUE(array->Equals(array));
-
-  // Equal, because the unequal index is masked by null
-  ASSERT_TRUE(array->Equals(array2));
-
-  // Unequal dictionaries
-  ASSERT_FALSE(array->Equals(array3));
-
-  // Unequal indices
-  ASSERT_FALSE(array->Equals(array4));
-
-  // RangeEquals
-  ASSERT_TRUE(array->RangeEquals(3, 6, 3, array4));
-  ASSERT_FALSE(array->RangeEquals(1, 3, 1, array4));
-
-  // ARROW-33 Test slices
-  const int64_t size = array->length();
-
-  std::shared_ptr<Array> slice, slice2;
-  slice = array->Array::Slice(2);
-  slice2 = array->Array::Slice(2);
-  ASSERT_EQ(size - 2, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(2, array->length(), 0, slice));
-
-  // Chained slices
-  slice2 = array->Array::Slice(1)->Array::Slice(1);
-  ASSERT_TRUE(slice->Equals(slice2));
-
-  slice = array->Slice(1, 3);
-  slice2 = array->Slice(1, 3);
-  ASSERT_EQ(3, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(1, 4, 0, slice));
-}
-
-TEST(TestDictionary, Validate) {
-  vector<bool> is_valid = {true, true, false, true, true, true};
-
-  std::shared_ptr<Array> dict;
-  vector<string> dict_values = {"foo", "bar", "baz"};
-  ArrayFromVector<StringType, string>(dict_values, &dict);
-  std::shared_ptr<DataType> dict_type = dictionary(int16(), dict);
-
-  std::shared_ptr<Array> indices;
-  vector<int16_t> indices_values = {1, 2, 0, 0, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(is_valid, indices_values, &indices);
-
-  std::shared_ptr<Array> arr = std::make_shared<DictionaryArray>(dict_type, indices);
-
-  // Only checking index type for now
-  ASSERT_OK(ValidateArray(*arr));
-
-  // TODO(wesm) In ARROW-1199, there is now a DCHECK to compare the indices
-  // type with the dict_type. How can we test for this?
-
-  // std::shared_ptr<Array> indices2;
-  // vector<float> indices2_values = {1., 2., 0., 0., 2., 0.};
-  // ArrayFromVector<FloatType, float>(is_valid, indices2_values, &indices2);
-
-  // std::shared_ptr<Array> indices3;
-  // vector<int64_t> indices3_values = {1, 2, 0, 0, 2, 0};
-  // ArrayFromVector<Int64Type, int64_t>(is_valid, indices3_values, &indices3);
-  // std::shared_ptr<Array> arr2 = std::make_shared<DictionaryArray>(dict_type, indices2);
-  // std::shared_ptr<Array> arr3 = std::make_shared<DictionaryArray>(dict_type, indices3);
-  // ASSERT_OK(ValidateArray(*arr3));
-}
-
-TEST(TestDictionary, FromArray) {
-  std::shared_ptr<Array> dict;
-  vector<string> dict_values = {"foo", "bar", "baz"};
-  ArrayFromVector<StringType, string>(dict_values, &dict);
-  std::shared_ptr<DataType> dict_type = dictionary(int16(), dict);
-
-  std::shared_ptr<Array> indices1;
-  vector<int16_t> indices_values1 = {1, 2, 0, 0, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(indices_values1, &indices1);
-
-  std::shared_ptr<Array> indices2;
-  vector<int16_t> indices_values2 = {1, 2, 0, 3, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(indices_values2, &indices2);
-
-  std::shared_ptr<Array> indices3;
-  vector<bool> is_valid3 = {true, true, false, true, true, true};
-  vector<int16_t> indices_values3 = {1, 2, -1, 0, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(is_valid3, indices_values3, &indices3);
-
-  std::shared_ptr<Array> indices4;
-  vector<bool> is_valid4 = {true, true, false, true, true, true};
-  vector<int16_t> indices_values4 = {1, 2, 1, 3, 2, 0};
-  ArrayFromVector<Int16Type, int16_t>(is_valid4, indices_values4, &indices4);
-
-  std::shared_ptr<Array> arr1, arr2, arr3, arr4;
-  ASSERT_OK(DictionaryArray::FromArrays(dict_type, indices1, &arr1));
-  ASSERT_RAISES(Invalid, DictionaryArray::FromArrays(dict_type, indices2, &arr2));
-  ASSERT_OK(DictionaryArray::FromArrays(dict_type, indices3, &arr3));
-  ASSERT_RAISES(Invalid, DictionaryArray::FromArrays(dict_type, indices4, &arr4));
-}
-
-// ----------------------------------------------------------------------
-// Struct tests
-
-void ValidateBasicStructArray(const StructArray* result,
-                              const vector<uint8_t>& struct_is_valid,
-                              const vector<char>& list_values,
-                              const vector<uint8_t>& list_is_valid,
-                              const vector<int>& list_lengths,
-                              const vector<int>& list_offsets,
-                              const vector<int32_t>& int_values) {
-  ASSERT_EQ(4, result->length());
-  ASSERT_OK(ValidateArray(*result));
-
-  auto list_char_arr = std::dynamic_pointer_cast<ListArray>(result->field(0));
-  auto char_arr = std::dynamic_pointer_cast<Int8Array>(list_char_arr->values());
-  auto int32_arr = std::dynamic_pointer_cast<Int32Array>(result->field(1));
-
-  ASSERT_EQ(nullptr, result->GetFieldByName("non-existing"));
-  ASSERT_TRUE(list_char_arr->Equals(result->GetFieldByName("list")));
-  ASSERT_TRUE(int32_arr->Equals(result->GetFieldByName("int")));
-
-  ASSERT_EQ(0, result->null_count());
-  ASSERT_EQ(1, list_char_arr->null_count());
-  ASSERT_EQ(0, int32_arr->null_count());
-
-  // List<char>
-  ASSERT_EQ(4, list_char_arr->length());
-  ASSERT_EQ(10, list_char_arr->values()->length());
-  for (size_t i = 0; i < list_offsets.size(); ++i) {
-    ASSERT_EQ(list_offsets[i], list_char_arr->raw_value_offsets()[i]);
-  }
-  for (size_t i = 0; i < list_values.size(); ++i) {
-    ASSERT_EQ(list_values[i], char_arr->Value(i));
-  }
-
-  // Int32
-  ASSERT_EQ(4, int32_arr->length());
-  for (size_t i = 0; i < int_values.size(); ++i) {
-    ASSERT_EQ(int_values[i], int32_arr->Value(i));
-  }
-}
-
-// ----------------------------------------------------------------------------------
-// Struct test
-class TestStructBuilder : public TestBuilder {
- public:
-  void SetUp() {
-    TestBuilder::SetUp();
-
-    auto int32_type = int32();
-    auto char_type = int8();
-    auto list_type = list(char_type);
-
-    vector<std::shared_ptr<DataType>> types = {list_type, int32_type};
-    vector<std::shared_ptr<Field>> fields;
-    fields.push_back(field("list", list_type));
-    fields.push_back(field("int", int32_type));
-
-    type_ = struct_(fields);
-    value_fields_ = fields;
-
-    std::unique_ptr<ArrayBuilder> tmp;
-    ASSERT_OK(MakeBuilder(pool_, type_, &tmp));
-    builder_.reset(checked_cast<StructBuilder*>(tmp.release()));
-    ASSERT_EQ(2, static_cast<int>(builder_->num_fields()));
-  }
-
-  void Done() {
-    std::shared_ptr<Array> out;
-    FinishAndCheckPadding(builder_.get(), &out);
-    result_ = std::dynamic_pointer_cast<StructArray>(out);
-  }
-
- protected:
-  vector<std::shared_ptr<Field>> value_fields_;
-
-  std::shared_ptr<StructBuilder> builder_;
-  std::shared_ptr<StructArray> result_;
-};
-
-TEST_F(TestStructBuilder, TestAppendNull) {
-  ASSERT_OK(builder_->AppendNull());
-  ASSERT_OK(builder_->AppendNull());
-  ASSERT_EQ(2, static_cast<int>(builder_->num_fields()));
-
-  ListBuilder* list_vb = checked_cast<ListBuilder*>(builder_->field_builder(0));
-  ASSERT_OK(list_vb->AppendNull());
-  ASSERT_OK(list_vb->AppendNull());
-  ASSERT_EQ(2, list_vb->length());
-
-  Int32Builder* int_vb = checked_cast<Int32Builder*>(builder_->field_builder(1));
-  ASSERT_OK(int_vb->AppendNull());
-  ASSERT_OK(int_vb->AppendNull());
-  ASSERT_EQ(2, int_vb->length());
-
-  Done();
-
-  ASSERT_OK(ValidateArray(*result_));
-
-  ASSERT_EQ(2, static_cast<int>(result_->num_fields()));
-  ASSERT_EQ(2, result_->length());
-  ASSERT_EQ(2, result_->field(0)->length());
-  ASSERT_EQ(2, result_->field(1)->length());
-  ASSERT_TRUE(result_->IsNull(0));
-  ASSERT_TRUE(result_->IsNull(1));
-  ASSERT_TRUE(result_->field(0)->IsNull(0));
-  ASSERT_TRUE(result_->field(0)->IsNull(1));
-  ASSERT_TRUE(result_->field(1)->IsNull(0));
-  ASSERT_TRUE(result_->field(1)->IsNull(1));
-
-  ASSERT_EQ(Type::LIST, result_->field(0)->type_id());
-  ASSERT_EQ(Type::INT32, result_->field(1)->type_id());
-}
-
-TEST_F(TestStructBuilder, TestBasics) {
-  vector<int32_t> int_values = {1, 2, 3, 4};
-  vector<char> list_values = {'j', 'o', 'e', 'b', 'o', 'b', 'm', 'a', 'r', 'k'};
-  vector<int> list_lengths = {3, 0, 3, 4};
-  vector<int> list_offsets = {0, 3, 3, 6, 10};
-  vector<uint8_t> list_is_valid = {1, 0, 1, 1};
-  vector<uint8_t> struct_is_valid = {1, 1, 1, 1};
-
-  ListBuilder* list_vb = checked_cast<ListBuilder*>(builder_->field_builder(0));
-  Int8Builder* char_vb = checked_cast<Int8Builder*>(list_vb->value_builder());
-  Int32Builder* int_vb = checked_cast<Int32Builder*>(builder_->field_builder(1));
-  ASSERT_EQ(2, static_cast<int>(builder_->num_fields()));
-
-  EXPECT_OK(builder_->Resize(list_lengths.size()));
-  EXPECT_OK(char_vb->Resize(list_values.size()));
-  EXPECT_OK(int_vb->Resize(int_values.size()));
-
-  int pos = 0;
-  for (size_t i = 0; i < list_lengths.size(); ++i) {
-    ASSERT_OK(list_vb->Append(list_is_valid[i] > 0));
-    int_vb->UnsafeAppend(int_values[i]);
-    for (int j = 0; j < list_lengths[i]; ++j) {
-      char_vb->UnsafeAppend(list_values[pos++]);
-    }
-  }
-
-  for (size_t i = 0; i < struct_is_valid.size(); ++i) {
-    ASSERT_OK(builder_->Append(struct_is_valid[i] > 0));
-  }
-
-  Done();
-
-  ValidateBasicStructArray(result_.get(), struct_is_valid, list_values, list_is_valid,
-                           list_lengths, list_offsets, int_values);
-}
-
-TEST_F(TestStructBuilder, BulkAppend) {
-  vector<int32_t> int_values = {1, 2, 3, 4};
-  vector<char> list_values = {'j', 'o', 'e', 'b', 'o', 'b', 'm', 'a', 'r', 'k'};
-  vector<int> list_lengths = {3, 0, 3, 4};
-  vector<int> list_offsets = {0, 3, 3, 6};
-  vector<uint8_t> list_is_valid = {1, 0, 1, 1};
-  vector<uint8_t> struct_is_valid = {1, 1, 1, 1};
-
-  ListBuilder* list_vb = checked_cast<ListBuilder*>(builder_->field_builder(0));
-  Int8Builder* char_vb = checked_cast<Int8Builder*>(list_vb->value_builder());
-  Int32Builder* int_vb = checked_cast<Int32Builder*>(builder_->field_builder(1));
-
-  ASSERT_OK(builder_->Resize(list_lengths.size()));
-  ASSERT_OK(char_vb->Resize(list_values.size()));
-  ASSERT_OK(int_vb->Resize(int_values.size()));
-
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  Done();
-  ValidateBasicStructArray(result_.get(), struct_is_valid, list_values, list_is_valid,
-                           list_lengths, list_offsets, int_values);
-}
-
-TEST_F(TestStructBuilder, BulkAppendInvalid) {
-  vector<int32_t> int_values = {1, 2, 3, 4};
-  vector<char> list_values = {'j', 'o', 'e', 'b', 'o', 'b', 'm', 'a', 'r', 'k'};
-  vector<int> list_lengths = {3, 0, 3, 4};
-  vector<int> list_offsets = {0, 3, 3, 6};
-  vector<uint8_t> list_is_valid = {1, 0, 1, 1};
-  vector<uint8_t> struct_is_valid = {1, 0, 1, 1};  // should be 1, 1, 1, 1
-
-  ListBuilder* list_vb = checked_cast<ListBuilder*>(builder_->field_builder(0));
-  Int8Builder* char_vb = checked_cast<Int8Builder*>(list_vb->value_builder());
-  Int32Builder* int_vb = checked_cast<Int32Builder*>(builder_->field_builder(1));
-
-  ASSERT_OK(builder_->Reserve(list_lengths.size()));
-  ASSERT_OK(char_vb->Reserve(list_values.size()));
-  ASSERT_OK(int_vb->Reserve(int_values.size()));
-
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  Done();
-  // Even null bitmap of the parent Struct is not valid, validate will ignore it.
-  ASSERT_OK(ValidateArray(*result_));
-}
-
-TEST_F(TestStructBuilder, TestEquality) {
-  std::shared_ptr<Array> array, equal_array;
-  std::shared_ptr<Array> unequal_bitmap_array, unequal_offsets_array,
-      unequal_values_array;
-
-  vector<int32_t> int_values = {101, 102, 103, 104};
-  vector<char> list_values = {'j', 'o', 'e', 'b', 'o', 'b', 'm', 'a', 'r', 'k'};
-  vector<int> list_lengths = {3, 0, 3, 4};
-  vector<int> list_offsets = {0, 3, 3, 6};
-  vector<uint8_t> list_is_valid = {1, 0, 1, 1};
-  vector<uint8_t> struct_is_valid = {1, 1, 1, 1};
-
-  vector<int32_t> unequal_int_values = {104, 102, 103, 101};
-  vector<char> unequal_list_values = {'j', 'o', 'e', 'b', 'o', 'b', 'l', 'u', 'c', 'y'};
-  vector<int> unequal_list_offsets = {0, 3, 4, 6};
-  vector<uint8_t> unequal_list_is_valid = {1, 1, 1, 1};
-  vector<uint8_t> unequal_struct_is_valid = {1, 0, 0, 1};
-
-  ListBuilder* list_vb = checked_cast<ListBuilder*>(builder_->field_builder(0));
-  Int8Builder* char_vb = checked_cast<Int8Builder*>(list_vb->value_builder());
-  Int32Builder* int_vb = checked_cast<Int32Builder*>(builder_->field_builder(1));
-  ASSERT_OK(builder_->Reserve(list_lengths.size()));
-  ASSERT_OK(char_vb->Reserve(list_values.size()));
-  ASSERT_OK(int_vb->Reserve(int_values.size()));
-
-  // setup two equal arrays, one of which takes an unequal bitmap
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  FinishAndCheckPadding(builder_.get(), &array);
-
-  ASSERT_OK(builder_->Resize(list_lengths.size()));
-  ASSERT_OK(char_vb->Resize(list_values.size()));
-  ASSERT_OK(int_vb->Resize(int_values.size()));
-
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  ASSERT_OK(builder_->Finish(&equal_array));
-
-  ASSERT_OK(builder_->Resize(list_lengths.size()));
-  ASSERT_OK(char_vb->Resize(list_values.size()));
-  ASSERT_OK(int_vb->Resize(int_values.size()));
-
-  // setup an unequal one with the unequal bitmap
-  ASSERT_OK(builder_->AppendValues(unequal_struct_is_valid.size(),
-                                   unequal_struct_is_valid.data()));
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  ASSERT_OK(builder_->Finish(&unequal_bitmap_array));
-
-  ASSERT_OK(builder_->Resize(list_lengths.size()));
-  ASSERT_OK(char_vb->Resize(list_values.size()));
-  ASSERT_OK(int_vb->Resize(int_values.size()));
-
-  // setup an unequal one with unequal offsets
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-  ASSERT_OK(list_vb->AppendValues(unequal_list_offsets.data(),
-                                  unequal_list_offsets.size(),
-                                  unequal_list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  ASSERT_OK(builder_->Finish(&unequal_offsets_array));
-
-  ASSERT_OK(builder_->Resize(list_lengths.size()));
-  ASSERT_OK(char_vb->Resize(list_values.size()));
-  ASSERT_OK(int_vb->Resize(int_values.size()));
-
-  // setup anunequal one with unequal values
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : unequal_list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : unequal_int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-
-  ASSERT_OK(builder_->Finish(&unequal_values_array));
-
-  // Test array equality
-  EXPECT_TRUE(array->Equals(array));
-  EXPECT_TRUE(array->Equals(equal_array));
-  EXPECT_TRUE(equal_array->Equals(array));
-  EXPECT_FALSE(equal_array->Equals(unequal_bitmap_array));
-  EXPECT_FALSE(unequal_bitmap_array->Equals(equal_array));
-  EXPECT_FALSE(unequal_bitmap_array->Equals(unequal_values_array));
-  EXPECT_FALSE(unequal_values_array->Equals(unequal_bitmap_array));
-  EXPECT_FALSE(unequal_bitmap_array->Equals(unequal_offsets_array));
-  EXPECT_FALSE(unequal_offsets_array->Equals(unequal_bitmap_array));
-
-  // Test range equality
-  EXPECT_TRUE(array->RangeEquals(0, 4, 0, equal_array));
-  EXPECT_TRUE(array->RangeEquals(3, 4, 3, unequal_bitmap_array));
-  EXPECT_TRUE(array->RangeEquals(0, 1, 0, unequal_offsets_array));
-  EXPECT_FALSE(array->RangeEquals(0, 2, 0, unequal_offsets_array));
-  EXPECT_FALSE(array->RangeEquals(1, 2, 1, unequal_offsets_array));
-  EXPECT_FALSE(array->RangeEquals(0, 1, 0, unequal_values_array));
-  EXPECT_TRUE(array->RangeEquals(1, 3, 1, unequal_values_array));
-  EXPECT_FALSE(array->RangeEquals(3, 4, 3, unequal_values_array));
-}
-
-TEST_F(TestStructBuilder, TestZeroLength) {
-  // All buffers are null
-  Done();
-  ASSERT_OK(ValidateArray(*result_));
-}
-
-TEST_F(TestStructBuilder, TestSlice) {
-  std::shared_ptr<Array> array, equal_array;
-  std::shared_ptr<Array> unequal_bitmap_array, unequal_offsets_array,
-      unequal_values_array;
-
-  vector<int32_t> int_values = {101, 102, 103, 104};
-  vector<char> list_values = {'j', 'o', 'e', 'b', 'o', 'b', 'm', 'a', 'r', 'k'};
-  vector<int> list_lengths = {3, 0, 3, 4};
-  vector<int> list_offsets = {0, 3, 3, 6};
-  vector<uint8_t> list_is_valid = {1, 0, 1, 1};
-  vector<uint8_t> struct_is_valid = {1, 1, 1, 1};
-
-  ListBuilder* list_vb = checked_cast<ListBuilder*>(builder_->field_builder(0));
-  Int8Builder* char_vb = checked_cast<Int8Builder*>(list_vb->value_builder());
-  Int32Builder* int_vb = checked_cast<Int32Builder*>(builder_->field_builder(1));
-  ASSERT_OK(builder_->Reserve(list_lengths.size()));
-  ASSERT_OK(char_vb->Reserve(list_values.size()));
-  ASSERT_OK(int_vb->Reserve(int_values.size()));
-
-  ASSERT_OK(builder_->AppendValues(struct_is_valid.size(), struct_is_valid.data()));
-  ASSERT_OK(list_vb->AppendValues(list_offsets.data(), list_offsets.size(),
-                                  list_is_valid.data()));
-  for (int8_t value : list_values) {
-    char_vb->UnsafeAppend(value);
-  }
-  for (int32_t value : int_values) {
-    int_vb->UnsafeAppend(value);
-  }
-  FinishAndCheckPadding(builder_.get(), &array);
-
-  std::shared_ptr<StructArray> slice, slice2;
-  std::shared_ptr<Int32Array> int_field;
-  std::shared_ptr<ListArray> list_field;
-
-  slice = std::dynamic_pointer_cast<StructArray>(array->Slice(2));
-  slice2 = std::dynamic_pointer_cast<StructArray>(array->Slice(2));
-  ASSERT_EQ(array->length() - 2, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(2, slice->length(), 0, slice));
-
-  int_field = std::dynamic_pointer_cast<Int32Array>(slice->field(1));
-  ASSERT_EQ(int_field->length(), slice->length());
-  ASSERT_EQ(int_field->Value(0), 103);
-  ASSERT_EQ(int_field->Value(1), 104);
-  ASSERT_EQ(int_field->null_count(), 0);
-  list_field = std::dynamic_pointer_cast<ListArray>(slice->field(0));
-  ASSERT_FALSE(list_field->IsNull(0));
-  ASSERT_FALSE(list_field->IsNull(1));
-  ASSERT_EQ(list_field->value_length(0), 3);
-  ASSERT_EQ(list_field->value_length(1), 4);
-  ASSERT_EQ(list_field->null_count(), 0);
-
-  slice = std::dynamic_pointer_cast<StructArray>(array->Slice(1, 2));
-  slice2 = std::dynamic_pointer_cast<StructArray>(array->Slice(1, 2));
-  ASSERT_EQ(2, slice->length());
-
-  ASSERT_TRUE(slice->Equals(slice2));
-  ASSERT_TRUE(array->RangeEquals(1, 3, 0, slice));
-
-  int_field = std::dynamic_pointer_cast<Int32Array>(slice->field(1));
-  ASSERT_EQ(int_field->length(), slice->length());
-  ASSERT_EQ(int_field->Value(0), 102);
-  ASSERT_EQ(int_field->Value(1), 103);
-  ASSERT_EQ(int_field->null_count(), 0);
-  list_field = std::dynamic_pointer_cast<ListArray>(slice->field(0));
-  ASSERT_TRUE(list_field->IsNull(0));
-  ASSERT_FALSE(list_field->IsNull(1));
-  ASSERT_EQ(list_field->value_length(0), 0);
-  ASSERT_EQ(list_field->value_length(1), 3);
-  ASSERT_EQ(list_field->null_count(), 1);
 }
 
 // ----------------------------------------------------------------------
